Add MediaWiki integration basics
